requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
distro==1.8.0
Flask==2.2.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
modern-treasury==1.11.0
pydantic==1.10.7
python-dotenv==0.21.1
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
urllib3==1.26.14
Werkzeug==2.2.2
